requirements.txt
gym
flask
flask-socketio
pillow
